backend/requirements.txt
asgiref==3.5.0
Django==4.0.1
django-extensions==3.1.5
djangorestframework==3.13.1
psycopg2-binary==2.9.3
pytz==2021.3
sqlparse==0.4.2
